+++ br.com.proway.senior.fechamentoFolha/src/main/java/br/com/proway/senior/controller/CalculosDesconto.java
@@ -1,4 +1,6 @@
-package br.com.proway.senior.model;
+package br.com.proway.senior.controller;
+
+import br.com.proway.senior.model.Folha;
 
 public class CalculosDesconto implements InterfaceImpostoDeRendaDesconto, InterfaceValeTransporteDesconto,
 		InterfaceINSSDesconto, InterfacePlanoDeSaudeDesconto {
