+++ br.com.proway.senior.fechamentoFolha/src/main/java/br/com/proway/senior/controller/InterfaceBonificacaoExtra.java
@@ -1,4 +1,4 @@
-package br.com.proway.senior.model;
+package br.com.proway.senior.controller;
 
 public interface InterfaceBonificacaoExtra {
