+++ br.com.proway.senior.fechamentoFolha/src/main/java/br/com/proway/senior/controller/InterfaceHorasFaltando.java
@@ -1,4 +1,4 @@
-package br.com.proway.senior.model;
+package br.com.proway.senior.controller;
 
 public interface InterfaceHorasFaltando {
 	public double calcularValorHorasFaltas();
