+++ br.com.proway.senior.fechamentoFolha/src/main/java/br/com/proway/senior/model/InterfaceFeriasFolha.java
@@ -1,0 +1,9 @@
+package br.com.proway.senior.model;
+
+public interface InterfaceFeriasFolha {
+
+	public int getDias(); 
+	
+	public int getAbono();
+	
+}
