+++ br.com.proway.senior.fechamentoFolha/src/main/java/br/com/proway/senior/model/InterfacePontoFolha.java
@@ -1,0 +1,11 @@
+package br.com.proway.senior.model;
+
+public interface InterfacePontoFolha {
+
+	public double getHorasTrabalhadas();
+	
+	public double getHoraExtra();
+	
+	public double getHorasFaltas();
+	
+}
