--- br.com.proway.senior.fechamentoFolha/src/main/java/br/com/proway/senior/model/Interfaces/InterfaceFerias.java
@@ -1,9 +1,0 @@
-package br.com.proway.senior.model.Interfaces;
-
-public interface InterfaceFerias {
-
-	public int getDias(); 
-	
-	public int getAbono();
-	
-}
