--- br.com.proway.senior.fechamentoFolha/src/main/java/br/com/proway/senior/model/Interfaces/InterfacePonto.java
@@ -1,11 +1,0 @@
-package br.com.proway.senior.model.Interfaces;
-
-public interface InterfacePonto {
-
-	public double getHorasTrablhadas();
-	
-	public double getHoraExtra();
-	
-	public double getHorasFaltas();
-	
-}
