+++ br.com.proway.senior.fechamentoFolha/src/main/java/br/com/proway/senior/model/PontoFolha.java
@@ -1,5 +1,5 @@
 package br.com.proway.senior.model;
 
-public class PontoFolha implements InterfacePonto{
+public class PontoFolha implements InterfacePontoFolha{
 
 }
